suite(chore): move buildCurrencyOptions to wallet-utils
