+++ packages/suite/src/views/wallet/coinmarket/exchange/components/ExchangeForm/Inputs/FiatInput/FiatSelect/index.tsx
@@ -1,23 +1,13 @@
 import { Select } from '@trezor/components';
 import React from 'react';
 import { Controller } from 'react-hook-form';
-import { FIAT } from '@suite-config';
 import { useCoinmarketExchangeFormContext } from '@wallet-hooks/useCoinmarketExchangeForm';
 import { FIAT_CURRENCY } from '@wallet-types/coinmarketExchangeForm';
-
-export const buildCurrencyOptions = () => {
-    const result: { value: string; label: string }[] = [];
-    FIAT.currencies.forEach(currency =>
-        result.push({ value: currency, label: currency.toUpperCase() }),
-    );
-
-    return result;
-};
+import { buildCurrencyOptions } from '@suite-common/wallet-utils';
 
 const FiatSelect = () => {
     const { control, setAmountLimits, updateFiatCurrency, defaultCurrency } =
         useCoinmarketExchangeFormContext();
-    const currencyOptions = buildCurrencyOptions();
 
     return (
         <Controller
@@ -33,7 +23,7 @@ const FiatSelect = () => {
                     }}
                     value={value}
                     isClearable={false}
-                    options={currencyOptions}
+                    options={buildCurrencyOptions(value)}
                     minWidth="58px"
                     isClean
                     hideTextCursor
